+++ utils.js
@@ -1,1 +1,10 @@
-export const randomNumber = (range) => Math.ceil(Math.random() * range)
+export const randomNumber = (range) => Math.ceil(Math.random() * range);
+
+export function createElement(tag, className) {
+  const $tag = document.createElement(tag);
+  if (className) {
+    $tag.classList.add(className);
+  }
+
+  return $tag;
+}
